- Active cases info txt change
Changed the algorithm that calculates the active cases. Since EODY decided to remove some data on the 12/02/21 the active cases had to re-figured from scratch. Now from the total cases of 17 days we subtract the total of foreign cases + total deaths. The percentage of foreign cases is not valid anymore since we dont have the data anymore.
Even that is not 100% accurate since the foreign cases will always be the same (last) number available to us...hopefully that will change in the future.